clean up recommended tags
final draft of recommended tags
