+++ .gitignore
@@ -49,3 +49,4 @@ ehthumbs_vista.db
 # Ionide folder, used in F# for VSCode
 .ionide/
 docs/core/extensions/snippets/workers/cloud-service/Properties/PublishProfiles/cloudworker.pubxml
+docs/csharp/language-reference/xmldoc/snippets/xmldoc/xmldoc.xml
