+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/c-tag.cs
@@ -1,0 +1,14 @@
+﻿namespace XmlTags
+{
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+        }
+}
