+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/exception-tag.cs
@@ -1,0 +1,79 @@
+﻿
+namespace ExceptionTag
+{
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+            /// <summary>
+        /// Adds two integers and returns the result.
+        /// </summary>
+        /// <returns>
+        /// The sum of two integers.
+        /// </returns>
+        /// <example>
+        /// <code>
+        /// int c = Math.Add(4, 5);
+        /// if (c > 10)
+        /// {
+        ///     Console.WriteLine(c);
+        /// }
+        /// </code>
+        /// </example>
+        /// <exception cref="System.OverflowException">Thrown when one parameter is max
+        /// and the other is greater than 0.</exception>
+            public static int Add(int a, int b)
+            {
+                if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+
+            /// <summary>
+        /// Adds two doubles and returns the result.
+        /// </summary>
+        /// <returns>
+        /// The sum of two doubles.
+        /// </returns>
+        /// <exception cref="System.OverflowException">Thrown when one parameter is max
+        /// and the other is greater than zero.</exception>
+            public static double Add(double a, double b)
+            {
+                if ((a == double.MaxValue && b > 0) || (b == double.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+
+            /// <summary>
+        /// Divides an integer by another and returns the result.
+        /// </summary>
+        /// <returns>
+        /// The division of two integers.
+        /// </returns>
+        /// <exception cref="System.DivideByZeroException">Thrown when a division by zero occurs.</exception>
+            public static int Divide(int a, int b)
+            {
+                return a / b;
+            }
+
+            /// <summary>
+        /// Divides a double by another and returns the result.
+        /// </summary>
+        /// <returns>
+        /// The division of two doubles.
+        /// </returns>
+        /// <exception cref="System.DivideByZeroException">Thrown when a division by zero occurs.</exception>
+            public static double Divide(double a, double b)
+            {
+                return a / b;
+            }
+        }
+}
