+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/include-tag.cs
@@ -1,0 +1,77 @@
+﻿namespace IncludeTag
+{
+
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <include file='docs.xml' path='docs/members[@name="math"]/Math/*'/>
+        public class Math
+        {
+            // Adds two integers and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/AddInt/*'/>
+            public static int Add(int a, int b)
+            {
+                // If any parameter is equal to the max value of an integer
+                // and the other is greater than zero
+                if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+
+            // Adds two doubles and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/AddDouble/*'/>
+            public static double Add(double a, double b)
+            {
+                // If any parameter is equal to the max value of an integer
+                // and the other is greater than zero
+                if ((a == double.MaxValue && b > 0) || (b == double.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+
+            // Subtracts an integer from another and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/SubtractInt/*'/>
+            public static int Subtract(int a, int b)
+            {
+                return a - b;
+            }
+
+            // Subtracts a double from another and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/SubtractDouble/*'/>
+            public static double Subtract(double a, double b)
+            {
+                return a - b;
+            }
+
+            // Multiplies two integers and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/MultiplyInt/*'/>
+            public static int Multiply(int a, int b)
+            {
+                return a * b;
+            }
+
+            // Multiplies two doubles and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/MultiplyDouble/*'/>
+            public static double Multiply(double a, double b)
+            {
+                return a * b;
+            }
+
+            // Divides an integer by another and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/DivideInt/*'/>
+            public static int Divide(int a, int b)
+            {
+                return a / b;
+            }
+
+            // Divides a double by another and returns the result
+            /// <include file='docs.xml' path='docs/members[@name="math"]/DivideDouble/*'/>
+            public static double Divide(double a, double b)
+            {
+                return a / b;
+            }
+        }
+}
