+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/inheritdoc-tag.cs
@@ -1,0 +1,20 @@
+﻿namespace InheritDoc
+{
+
+    /*
+        The IMath interface
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// This is the IMath interface.
+    /// </summary>
+    public interface IMath
+        {
+        }
+
+        /// <inheritdoc/>
+        public class Math : IMath
+        {
+        }
+}
