+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/list-tag.cs
@@ -1,0 +1,33 @@
+﻿namespace ListTag
+{
+    /*
+    The main Math class
+    Contains all methods for performing basic math functions
+*/
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// <list type="bullet">
+    /// <item>
+    /// <term>Add</term>
+    /// <description>Addition Operation</description>
+    /// </item>
+    /// <item>
+    /// <term>Subtract</term>
+    /// <description>Subtraction Operation</description>
+    /// </item>
+    /// <item>
+    /// <term>Multiply</term>
+    /// <description>Multiplication Operation</description>
+    /// </item>
+    /// <item>
+    /// <term>Divide</term>
+    /// <description>Division Operation</description>
+    /// </item>
+    /// </list>
+    /// </summary>
+    public class Math
+        {
+        }
+
+}
