+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/para-tag.cs
@@ -1,0 +1,18 @@
+﻿namespace ParaTag
+{
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// The main Math class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    /// <remarks>
+    /// <para>This class can add, subtract, multiply and divide.</para>
+    /// <para>These operations can be performed on both integers and doubles.</para>
+    /// </remarks>
+    public class Math
+        {
+        }
+}
