+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/param-tag.cs
@@ -1,0 +1,32 @@
+﻿namespace ParamTag
+{
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+            /// <summary>
+        /// Adds two doubles and returns the result.
+        /// </summary>
+        /// <returns>
+        /// The sum of two doubles.
+        /// </returns>
+        /// <exception cref="System.OverflowException">Thrown when one parameter is max
+        /// and the other is greater than zero.</exception>
+        /// See <see cref="Math.Add(int, int)"/> to add integers.
+        /// <param name="a">A double precision number.</param>
+        /// <param name="b">A double precision number.</param>
+            public static double Add(double a, double b)
+            {
+                if ((a == double.MaxValue && b > 0) || (b == double.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+        }
+}
