+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/remarks-tag.cs
@@ -1,0 +1,17 @@
+﻿namespace RemarksTag
+{
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// The main Math class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    /// <remarks>
+    /// This class can add, subtract, multiply and divide.
+    /// </remarks>
+    public class Math
+        {
+        }
+}
