+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/seealso-tag.cs
@@ -1,0 +1,42 @@
+﻿namespace SeeAlsoTag
+{
+    /*
+    The main Math class
+    Contains all methods for performing basic math functions
+*/
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+            /// <summary>
+        /// Adds two integers and returns the result.
+        /// </summary>
+        /// <returns>
+        /// The sum of two integers.
+        /// </returns>
+        /// <example>
+        /// <code>
+        /// int c = Math.Add(4, 5);
+        /// if (c > 10)
+        /// {
+        ///     Console.WriteLine(c);
+        /// }
+        /// </code>
+        /// </example>
+        /// <exception cref="System.OverflowException">Thrown when one parameter is max
+        /// and the other is greater than 0.</exception>
+        /// See <see cref="Math.Add(double, double)"/> to add doubles.
+        /// <seealso cref="Math.Subtract(int, int)"/>
+        /// <seealso cref="Math.Multiply(int, int)"/>
+        /// <seealso cref="Math.Divide(int, int)"/>
+            public static int Add(int a, int b)
+            {
+                if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+        }
+}
