+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/summary-tag.cs
@@ -1,0 +1,28 @@
+﻿
+namespace SummaryTag
+{
+    /*
+    The main Math class
+    Contains all methods for performing basic math functions
+*/
+        /// <summary>
+    /// The main Math class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+            // Adds two integers and returns the result
+            /// <summary>
+        /// Adds two integers and returns the result.
+        /// </summary>
+            public static int Add(int a, int b)
+            {
+                // If any parameter is equal to the max value of an integer
+                // and the other is greater than zero
+                if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
+                    throw new System.OverflowException();
+
+                return a + b;
+            }
+        }
+}
