+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/typeparam-tag.cs
@@ -1,0 +1,25 @@
+﻿
+using System;
+
+namespace TypeParamLibrary
+{
+    /*
+    The main Math class
+    Contains all methods for performing basic math functions
+*/
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+            /// <summary>
+        /// Checks if an IComparable is greater than another.
+        /// </summary>
+        /// <typeparam name="T">A type that inherits from the IComparable interface.</typeparam>
+            public static bool GreaterThan<T>(T a, T b) where T : IComparable
+            {
+                return a.CompareTo(b) > 0;
+            }
+        }
+}
