+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/typeparamref-tag.cs
@@ -1,0 +1,24 @@
+﻿using System;
+
+namespace TypeParamRef
+{
+    /*
+    The main Math class
+    Contains all methods for performing basic math functions
+*/
+        /// <summary>
+    /// The main <c>Math</c> class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    public class Math
+        {
+            /// <summary>
+        /// Checks if an IComparable <typeparamref name="T"/> is greater than another.
+        /// </summary>
+        /// <typeparam name="T">A type that inherits from the IComparable interface.</typeparam>
+            public static bool GreaterThan<T>(T a, T b) where T : IComparable
+            {
+                return a.CompareTo(b) > 0;
+            }
+        }
+}
