+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/value-tag.cs
@@ -1,0 +1,20 @@
+﻿namespace ValueTag
+{
+    /*
+        The main Math class
+        Contains all methods for performing basic math functions
+    */
+        /// <summary>
+    /// The main Math class.
+    /// Contains all methods for performing basic math functions.
+    /// </summary>
+    /// <remarks>
+    /// This class can add, subtract, multiply and divide.
+    /// These operations can be performed on both integers and doubles
+    /// </remarks>
+    public class Math
+        {
+            /// <value>Gets the value of PI.</value>
+            public static double PI { get; }
+        }
+}
