+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/xml_include_tag.xml
@@ -1,0 +1,12 @@
+<MyDocs>
+    <MyMembers name="test">
+        <summary>
+        The summary for this type.
+        </summary>
+    </MyMembers>
+    <MyMembers name="test2">
+        <summary>
+        The summary for this other type.
+        </summary>
+    </MyMembers>
+</MyDocs>
