--- samples/snippets/csharp/concepts/codedoc/inheritdoc-tag.cs
@@ -1,16 +1,0 @@
-﻿/*
-    The IMath interface
-    The main Math class
-    Contains all methods for performing basic math functions
-*/
-/// <summary>
-/// This is the IMath interface.
-/// </summary>
-public interface IMath
-{
-}
-
-/// <inheritdoc/>
-public class Math : IMath
-{
-}
