--- samples/snippets/csharp/concepts/codedoc/remarks-tag.cs
@@ -1,14 +1,0 @@
-﻿/*
-    The main Math class
-    Contains all methods for performing basic math functions
-*/
-/// <summary>
-/// The main Math class.
-/// Contains all methods for performing basic math functions.
-/// </summary>
-/// <remarks>
-/// This class can add, subtract, multiply and divide.
-/// </remarks>
-public class Math
-{
-}
