--- samples/snippets/csharp/concepts/codedoc/returns-tag.cs
@@ -1,16 +1,0 @@
-﻿// Adds two integers and returns the result
-/// <summary>
-/// Adds two integers and returns the result.
-/// </summary>
-/// <returns>
-/// The sum of two integers.
-/// </returns>
-public static int Add(int a, int b)
-{
-    // If any parameter is equal to the max value of an integer
-    // and the other is greater than zero
-    if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
-        throw new System.OverflowException();
-
-    return a + b;
-}
