--- samples/snippets/csharp/concepts/codedoc/see-tag.cs
@@ -1,53 +1,0 @@
-﻿/*
-    The main Math class
-    Contains all methods for performing basic math functions
-*/
-/// <summary>
-/// The main <c>Math</c> class.
-/// Contains all methods for performing basic math functions.
-/// </summary>
-public class Math
-{
-    /// <summary>
-    /// Adds two integers and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The sum of two integers.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// int c = Math.Add(4, 5);
-    /// if (c > 10)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// <exception cref="System.OverflowException">Thrown when one parameter is max
-    /// and the other is greater than 0.</exception>
-    /// See <see cref="Math.Add(double, double)"/> to add doubles.
-    public static int Add(int a, int b)
-    {
-        if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
-            throw new System.OverflowException();
-
-        return a + b;
-    }
-
-    /// <summary>
-    /// Adds two doubles and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The sum of two doubles.
-    /// </returns>
-    /// <exception cref="System.OverflowException">Thrown when one parameter is max
-    /// and the other is greater than zero.</exception>
-    /// See <see cref="Math.Add(int, int)"/> to add integers.
-    public static double Add(double a, double b)
-    {
-        if ((a == double.MaxValue && b > 0) || (b == double.MaxValue && a > 0))
-            throw new System.OverflowException();
-
-        return a + b;
-    }
-}
