--- samples/snippets/csharp/concepts/codedoc/tagged-library.cs
@@ -1,264 +1,0 @@
-﻿/*
-    The main Math class
-    Contains all methods for performing basic math functions
-*/
-/// <summary>
-/// The main <c>Math</c> class.
-/// Contains all methods for performing basic math functions.
-/// <list type="bullet">
-/// <item>
-/// <term>Add</term>
-/// <description>Addition Operation</description>
-/// </item>
-/// <item>
-/// <term>Subtract</term>
-/// <description>Subtraction Operation</description>
-/// </item>
-/// <item>
-/// <term>Multiply</term>
-/// <description>Multiplication Operation</description>
-/// </item>
-/// <item>
-/// <term>Divide</term>
-/// <description>Division Operation</description>
-/// </item>
-/// </list>
-/// </summary>
-/// <remarks>
-/// <para>This class can add, subtract, multiply and divide.</para>
-/// <para>These operations can be performed on both integers and doubles.</para>
-/// </remarks>
-public class Math
-{
-    // Adds two integers and returns the result
-    /// <summary>
-    /// Adds two integers <paramref name="a"/> and <paramref name="b"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The sum of two integers.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// int c = Math.Add(4, 5);
-    /// if (c > 10)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// <exception cref="System.OverflowException">Thrown when one parameter is max
-    /// and the other is greater than 0.</exception>
-    /// See <see cref="Math.Add(double, double)"/> to add doubles.
-    /// <seealso cref="Math.Subtract(int, int)"/>
-    /// <seealso cref="Math.Multiply(int, int)"/>
-    /// <seealso cref="Math.Divide(int, int)"/>
-    /// <param name="a">An integer.</param>
-    /// <param name="b">An integer.</param>
-    public static int Add(int a, int b)
-    {
-        // If any parameter is equal to the max value of an integer
-        // and the other is greater than zero
-        if ((a == int.MaxValue && b > 0) || (b == int.MaxValue && a > 0))
-            throw new System.OverflowException();
-
-        return a + b;
-    }
-
-    // Adds two doubles and returns the result
-    /// <summary>
-    /// Adds two doubles <paramref name="a"/> and <paramref name="b"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The sum of two doubles.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// double c = Math.Add(4.5, 5.4);
-    /// if (c > 10)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// <exception cref="System.OverflowException">Thrown when one parameter is max
-    /// and the other is greater than 0.</exception>
-    /// See <see cref="Math.Add(int, int)"/> to add integers.
-    /// <seealso cref="Math.Subtract(double, double)"/>
-    /// <seealso cref="Math.Multiply(double, double)"/>
-    /// <seealso cref="Math.Divide(double, double)"/>
-    /// <param name="a">A double precision number.</param>
-    /// <param name="b">A double precision number.</param>
-    public static double Add(double a, double b)
-    {
-        // If any parameter is equal to the max value of an integer
-        // and the other is greater than zero
-        if ((a == double.MaxValue && b > 0) || (b == double.MaxValue && a > 0))
-            throw new System.OverflowException();
-
-        return a + b;
-    }
-
-    // Subtracts an integer from another and returns the result
-    /// <summary>
-    /// Subtracts <paramref name="b"/> from <paramref name="a"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The difference between two integers.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// int c = Math.Subtract(4, 5);
-    /// if (c > 1)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// See <see cref="Math.Subtract(double, double)"/> to subtract doubles.
-    /// <seealso cref="Math.Add(int, int)"/>
-    /// <seealso cref="Math.Multiply(int, int)"/>
-    /// <seealso cref="Math.Divide(int, int)"/>
-    /// <param name="a">An integer.</param>
-    /// <param name="b">An integer.</param>
-    public static int Subtract(int a, int b)
-    {
-        return a - b;
-    }
-
-    // Subtracts a double from another and returns the result
-    /// <summary>
-    /// Subtracts a double <paramref name="b"/> from another double <paramref name="a"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The difference between two doubles.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// double c = Math.Subtract(4.5, 5.4);
-    /// if (c > 1)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// See <see cref="Math.Subtract(int, int)"/> to subtract integers.
-    /// <seealso cref="Math.Add(double, double)"/>
-    /// <seealso cref="Math.Multiply(double, double)"/>
-    /// <seealso cref="Math.Divide(double, double)"/>
-    /// <param name="a">A double precision number.</param>
-    /// <param name="b">A double precision number.</param>
-    public static double Subtract(double a, double b)
-    {
-        return a - b;
-    }
-
-    // Multiplies two integers and returns the result
-    /// <summary>
-    /// Multiplies two integers <paramref name="a"/> and <paramref name="b"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The product of two integers.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// int c = Math.Multiply(4, 5);
-    /// if (c > 100)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// See <see cref="Math.Multiply(double, double)"/> to multiply doubles.
-    /// <seealso cref="Math.Add(int, int)"/>
-    /// <seealso cref="Math.Subtract(int, int)"/>
-    /// <seealso cref="Math.Divide(int, int)"/>
-    /// <param name="a">An integer.</param>
-    /// <param name="b">An integer.</param>
-    public static int Multiply(int a, int b)
-    {
-        return a * b;
-    }
-
-    // Multiplies two doubles and returns the result
-    /// <summary>
-    /// Multiplies two doubles <paramref name="a"/> and <paramref name="b"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The product of two doubles.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// double c = Math.Multiply(4.5, 5.4);
-    /// if (c > 100.0)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// See <see cref="Math.Multiply(int, int)"/> to multiply integers.
-    /// <seealso cref="Math.Add(double, double)"/>
-    /// <seealso cref="Math.Subtract(double, double)"/>
-    /// <seealso cref="Math.Divide(double, double)"/>
-    /// <param name="a">A double precision number.</param>
-    /// <param name="b">A double precision number.</param>
-    public static double Multiply(double a, double b)
-    {
-        return a * b;
-    }
-
-    // Divides an integer by another and returns the result
-    /// <summary>
-    /// Divides an integer <paramref name="a"/> by another integer <paramref name="b"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The quotient of two integers.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// int c = Math.Divide(4, 5);
-    /// if (c > 1)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// <exception cref="System.DivideByZeroException">Thrown when <paramref name="b"/> is equal to 0.</exception>
-    /// See <see cref="Math.Divide(double, double)"/> to divide doubles.
-    /// <seealso cref="Math.Add(int, int)"/>
-    /// <seealso cref="Math.Subtract(int, int)"/>
-    /// <seealso cref="Math.Multiply(int, int)"/>
-    /// <param name="a">An integer dividend.</param>
-    /// <param name="b">An integer divisor.</param>
-    public static int Divide(int a, int b)
-    {
-        return a / b;
-    }
-
-    // Divides a double by another and returns the result
-    /// <summary>
-    /// Divides a double <paramref name="a"/> by another double <paramref name="b"/> and returns the result.
-    /// </summary>
-    /// <returns>
-    /// The quotient of two doubles.
-    /// </returns>
-    /// <example>
-    /// <code>
-    /// double c = Math.Divide(4.5, 5.4);
-    /// if (c > 1.0)
-    /// {
-    ///     Console.WriteLine(c);
-    /// }
-    /// </code>
-    /// </example>
-    /// <exception cref="System.DivideByZeroException">Thrown when <paramref name="b"/> is equal to 0.</exception>
-    /// See <see cref="Math.Divide(int, int)"/> to divide integers.
-    /// <seealso cref="Math.Add(double, double)"/>
-    /// <seealso cref="Math.Subtract(double, double)"/>
-    /// <seealso cref="Math.Multiply(double, double)"/>
-    /// <param name="a">A double precision dividend.</param>
-    /// <param name="b">A double precision divisor.</param>
-    public static double Divide(double a, double b)
-    {
-        return a / b;
-    }
-}
